app/requirements.txt
streamlit>=1.28.0
strands-agents>=0.2.0
boto3>=1.34.0
httpx>=0.27.0
pytz>=2023.3
mcp>=1.0.0 